refactor: manage yt-dlp and ffmpeg binaries
